// src/shared/services/api/models/leaderboardDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface LeaderboardDto {
  user_id?: string;
  /** @nullable */
  full_name?: string | null;
  overall_weighted_score?: number;
  weighted_score?: number;
  rank?: number;
}